implement correct reading, writing and deleting of indexes.

+++ collection.go
@@ -10,6 +10,7 @@ import (
 	"log"
 	"os"
 	"path/filepath"
+	"strconv"
 )
 
 type Collection struct {
@@ -68,7 +69,7 @@ func (c *Collection) loadIndexes() {
 }
 
 func (c *Collection) loadIndex(filepath, field string) error {
-	file, err := os.Open(filepath)
+	file, err := os.OpenFile(filepath, os.O_RDWR, 0644)
 	if err != nil {
 		return err
 	}
@@ -115,12 +116,18 @@ func (c *Collection) Insert(value interface{}) (Id, error) {
 	}
 
 	id := c.getNextId()
-	err = c.store.Write(fmt.Sprintf("%d", id), jsondata)
+	id_str := fmt.Sprintf("%d", id)
+	err = c.store.Write(id_str, jsondata)
 	if err != nil {
 		c.setNextId(id) // roll back generated ID
-		id = Id(0)      // set id to invalid value
+		return Id(0), err
 	}
-	return id, err
+
+	if err = c.addToIndexes(id, jsondata); err != nil {
+		c.store.Erase(id_str)
+		return Id(0), err
+	}
+	return id, nil
 }
 
 func (c *Collection) Update(id Id, value interface{}) error {
@@ -129,17 +136,97 @@ func (c *Collection) Update(id Id, value interface{}) error {
 		return err
 	}
 
-	return c.store.Write(fmt.Sprintf("%d", id), jsondata)
+	if err = c.store.Write(fmt.Sprintf("%d", id), jsondata); err != nil {
+		return err
+	}
+
+	c.removeFromIndexes(id)
+
+	if err = c.addToIndexes(id, jsondata); err != nil {
+		return err
+	}
+	return nil
+}
+
+func (c *Collection) addToIndexes(id Id, jsondata []byte) error {
+	var value2 map[string]interface{}
+	// no error means that we can unmarshal it into a map.
+	if err := json.Unmarshal(jsondata, &value2); err == nil {
+		for field, idx := range c.indexes {
+			if v, contains := value2[field]; contains {
+				entry := indexEntry{deleted: false, value: fmt.Sprintf("%v", v), id: int64(id)}
+				fpos, _ := idx.file.Seek(0, os.SEEK_END)
+				if _, err = entry.WriteTo(idx.file); err != nil {
+					return err
+				}
+				idx.file.Sync()
+				entry.fpos = fpos
+				idx.Add(entry)
+			}
+		}
+	}
+	return nil
 }
 
 func (c *Collection) AddIndex(field string) error {
-	return errors.New("adding index failed")
+	filepath := c.indexpath + "/" + field
+
+	file, err := os.OpenFile(filepath, os.O_WRONLY | os.O_CREATE | os.O_EXCL, 0644)
+	if err != nil {
+		// if we couldn't open the file because it already exists, then AddIndex is a no-op.
+		if os.IsExist(err) {
+			return nil
+		}
+		return err
+	}
+
+	idx := newIndex(file, field)
+
+	for id_str := range c.store.Keys() {
+		id, err := strconv.ParseInt(id_str, 10, 64)
+		if err != nil {
+			log.Printf("AddIndex: skipping key %s", id_str)
+			continue
+		}
+
+		var entry map[string]interface{}
+		data, err := c.store.Read(id_str)
+		if err != nil {
+			log.Printf("AddIndex: skipping key %s because read from store failed: %v", id_str, err)
+			continue
+		}
+
+		if err = json.Unmarshal(data, &entry); err != nil {
+			log.Printf("AddIndex: skipping key %s because unmarshaling failed: %v", id_str, err)
+			continue
+		}
+
+		if value, exists := entry[field]; exists {
+			entry := indexEntry{deleted: false, value: fmt.Sprintf("%v", value), id: id}
+			fpos, _ := file.Seek(0, os.SEEK_END)
+			if _, err := entry.WriteTo(file); err != nil {
+				log.Printf("AddIndex: writing to index file failed: %v", err)
+				c.RemoveIndex(field)
+				return err
+			}
+			file.Sync()
+			entry.fpos = fpos
+			idx.Add(entry)
+		}
+	}
+
+	c.indexes[field] = idx
+
+	return nil
 }
 
 func (c *Collection) RemoveIndex(field string) error {
-	delete(c.indexes, field)
-	if err := os.Remove(c.indexpath + "/" + field); err != nil {
-		return err
+	if idx, exists := c.indexes[field]; exists {
+		idx.file.Close()
+		delete(c.indexes, field)
+		if err := os.Remove(c.indexpath + "/" + field); err != nil {
+			return err
+		}
 	}
 	return nil
 }
@@ -151,7 +238,32 @@ func (c *Collection) Reindex(field string) error {
 	return c.AddIndex(field)
 }
 
+func (c *Collection) removeFromIndexes(id Id) {
+	// remove entries from indexes
+	for _, idx := range c.indexes {
+		for key, entries := range idx.data {
+			new_entries := []indexEntry{}
+			for _, e := range entries {
+				if e.id != int64(id) {
+					new_entries = append(new_entries, e)
+				} else {
+					idx.file.Seek(e.fpos, os.SEEK_SET)
+					e.deleted = true
+					e.WriteTo(idx.file)
+					idx.file.Sync()
+				}
+			}
+			if len(new_entries) > 0 {
+				idx.data[key] = new_entries
+			} else {
+				delete(idx.data, key)
+			}
+		}
+	}
+}
+
 func (c *Collection) Delete(id Id) error {
+	c.removeFromIndexes(id)
 	return c.store.Erase(fmt.Sprintf("%d", id))
 }
 
@@ -164,6 +276,46 @@ func (c *Collection) QueryAll() (*Result, error) {
 }
 
 func (c *Collection) Vacuum() error {
-	// TODO: clean up indexes
+	for field, _ := range c.indexes {
+		oldf, err := os.Open(c.indexpath + "/" + field)
+		if err != nil {
+			return err
+		}
+		defer oldf.Close()
+
+		newf, err := os.OpenFile(c.indexpath + "/." + field + ".tmp", os.O_WRONLY | os.O_CREATE, 0644)
+		if err != nil {
+			return err
+		}
+		defer newf.Close()
+
+		for {
+			var entry indexEntry
+			n, err := entry.ReadFrom(oldf)
+			if err != nil {
+				if err == io.EOF {
+					break
+				}
+				return err
+			}
+
+			// don't write deleted entries
+			if entry.Deleted() {
+				continue
+			}
+
+			m, err := entry.WriteTo(newf)
+			if err != nil {
+				return err
+			}
+			if n != m {
+				os.Remove(c.indexpath + "/." + field + ".tmp")
+				return fmt.Errorf("short write while writing new index for %s", field)
+			}
+		}
+
+		os.Rename(c.indexpath + "/." + field + ".tmp", c.indexpath + "/" + field)
+		newf.Sync()
+	}
 	return nil
 }
